requirements.txt
# Streamlit for UI
streamlit==1.25.0
pillow==10.0.0
imagehash==4.3.1
numpy==1.26.0
openai==1.3.0
twilio==8.4.1
python-dotenv==1.0.0
scikit-learn==1.2.2
camel-ai==0.2.5
